app/tools: extract catalog IO definitions and tests
